+++ src/kOS.Safe/Execution/CPU.cs
@@ -70,6 +70,12 @@ namespace kOS.Safe.Execution
         /// <value><c>true</c> if this CPU is popping context; otherwise, <c>false</c>.</value>
         public bool IsPoppingContext { get; private set; }
 
+        /// <summary>
+        /// The objects which have chosen to register themselves as IPopContextNotifyees
+        /// to be told when popping a context (ending a program).
+        /// </summary>
+        private List<WeakReference> popContextNotifyees;
+
         public CPU(SafeSharedObjects shared)
         {
             this.shared = shared;
@@ -80,6 +86,7 @@ namespace kOS.Safe.Execution
             mainYields = new List<YieldFinishedDetector>();
             triggerYields = new List<YieldFinishedDetector>();
             if (this.shared.UpdateHandler != null) this.shared.UpdateHandler.AddFixedObserver(this);
+            popContextNotifyees = new List<WeakReference>();
         }
 
         public void Boot()
@@ -201,6 +208,7 @@ namespace kOS.Safe.Execution
             {
                 // remove the last context
                 ProgramContext contextRemove = contexts.Last();
+                NotifyPopContextNotifyees(contextRemove);
                 contexts.Remove(contextRemove);
                 shared.GameEventDispatchManager.RemoveDispatcherFor(currentContext);
                 contextRemove.DisableActiveFlyByWire(shared.BindingMgr);
@@ -226,6 +234,54 @@ namespace kOS.Safe.Execution
                 }
             }
             IsPoppingContext = false;
+        }
+
+        /// <summary>
+        /// Used when an object wants the CPU to call its OnPopContext() callback
+        /// whenever the CPU ends a program context.  Notice that the CPU will
+        /// only use a WEAK refernece to store this, so that registering yourself
+        /// as a notifyee here does not stop you from being orphaned and garbage
+        /// collected if you would normally do so.  It's only if your object happens
+        /// to still be alive when the program context ends that you'll get called
+        /// by the CPU.  Use this if you have some important cleanup work you'd like
+        /// to do when the program dies, or if your object is one that would become
+        /// useless and invalid when the program context ends so you need to shut
+        /// yourself down when that happens.
+        /// </summary>
+        /// <param name="notifyee">Notifyee object that has an OnPopContext() callback.</param>
+        public void AddPopContextNotifyee(IPopContextNotifyee notifyee)
+        {
+            // Not sure what the definition of Equals is for a weak reference,
+            // this walks through looking if it's already registered, to avoid duplicates:
+            for (int i = 0; i < popContextNotifyees.Count; ++i)
+                if (popContextNotifyees[i].Target == notifyee)
+                    return;
+
+            popContextNotifyees.Add(new WeakReference(notifyee));
+        }
+
+        public void RemovePopContextNotifyee(IPopContextNotifyee notifyee)
+        {
+            // Might as well also get rid of any that are stale references while we're here:
+            popContextNotifyees.RemoveAll((item)=>(!item.IsAlive) || item.Target == notifyee);
+        }
+
+        private void NotifyPopContextNotifyees(IProgramContext context)
+        {
+            // Notify them all:
+            for (int i = 0; i < popContextNotifyees.Count; ++i)
+            {
+                WeakReference current = popContextNotifyees[i];
+                if (current.IsAlive) // Avoid resurrecting it if it's gone, and don't call its hook.
+                {
+                    IPopContextNotifyee notifyee = current.Target as IPopContextNotifyee;
+                    if (!notifyee.OnPopContext(context))
+                        current.Target = null; // mark for removal below, because the notifyee wants us to
+                }
+            }
+
+            // Remove the ones flagged for removal or that are stale anyway:
+            popContextNotifyees.RemoveAll((item)=>(!item.IsAlive) || item.Target == null);
         }
 
         /// <summary>
@@ -753,7 +809,7 @@ namespace kOS.Safe.Execution
             if (barewordOkay)
             {
                 string strippedIdent = identifier.TrimStart('$');
-                return new Variable { Name = strippedIdent, Value = strippedIdent };
+                return new Variable { Name = strippedIdent, Value = new StringValue(strippedIdent) };
             }
             if (failOkay)
                 return null;
@@ -1106,11 +1162,13 @@ namespace kOS.Safe.Execution
         /// when it finishes.  If its return is false, it won't fire off again.
         /// </summary>
         /// <param name="triggerFunctionPointer">The entry point of this trigger function.</param>
+        /// <param name="closure">The closure the trigger should be called with.  If this is
+        /// null, then the trigger will only be able to see global variables reliably.</param>
         /// <returns>A TriggerInfo structure describing this new trigger, which probably isn't very useful
         /// tp the caller in most circumstances where this is a fire-and-forget trigger.</returns>
-        public TriggerInfo AddTrigger(int triggerFunctionPointer)
+        public TriggerInfo AddTrigger(int triggerFunctionPointer, List<VariableScope> closure)
         {
-            TriggerInfo triggerRef = new TriggerInfo(currentContext, triggerFunctionPointer);
+            TriggerInfo triggerRef = new TriggerInfo(currentContext, triggerFunctionPointer, closure);
             currentContext.AddPendingTrigger(triggerRef);
             return triggerRef;
         }
@@ -1144,7 +1202,7 @@ namespace kOS.Safe.Execution
         {
             if (del.ProgContext != currentContext)
                 return null;
-            TriggerInfo callbackRef = new TriggerInfo(currentContext, del.EntryPoint, args);
+            TriggerInfo callbackRef = new TriggerInfo(currentContext, del.EntryPoint, del.Closure, args);
             currentContext.AddPendingTrigger(callbackRef);
             return callbackRef;
         }
@@ -1206,7 +1264,7 @@ namespace kOS.Safe.Execution
 
         public void RemoveTrigger(int triggerFunctionPointer)
         {
-            currentContext.RemoveTrigger(new TriggerInfo(currentContext, triggerFunctionPointer));
+            currentContext.RemoveTrigger(new TriggerInfo(currentContext, triggerFunctionPointer, null));
         }
 
         public void RemoveTrigger(TriggerInfo trigger)
@@ -1357,12 +1415,15 @@ namespace kOS.Safe.Execution
                         // Insert a faked function call as if the trigger had been called from just
                         // before whatever opcode was about to get executed, by pusing a context
                         // record like OpcodeCall would do, and moving the IP to the
-                        // first line of the trigger, like OpcodeCall would do.  Because
-                        // triggers can't take arguments, most of the messy work of
-                        // OpcodeCall.Execute isn't needed:
+                        // first line of the trigger, like OpcodeCall would do.
                         SubroutineContext contextRecord =
                             new SubroutineContext(currentInstructionPointer, trigger);
                         PushAboveStack(contextRecord);
+
+                        // Reverse-push the closure's scope record, if there is one, just after the function return context got put on the stack.
+                        if (trigger.Closure != null)
+                            for (int i = trigger.Closure.Count - 1 ; i >= 0 ; --i)
+                                PushAboveStack(trigger.Closure[i]);
 
                         PushStack(new KOSArgMarkerType());
 
@@ -1371,7 +1432,7 @@ namespace kOS.Safe.Execution
                                 PushStack(trigger.Args[argIndex]);
                         
                         triggersToBeExecuted.Add(trigger);
-                        
+
                         currentInstructionPointer = trigger.EntryPoint;
                         // Triggers can chain in this loop if more than one fire off at once - the second trigger
                         // will look like it was a function that was called from the start of the first trigger.
@@ -1580,6 +1641,11 @@ namespace kOS.Safe.Execution
 
         public void Dispose()
         {
+            while (contexts.Count > 0)
+            {
+                PopContext();
+            }
+            contexts.Clear();
             shared.UpdateHandler.RemoveFixedObserver(this);
         }