+++ src/kOS/Module/SceneChangeCleaner.cs
@@ -1,0 +1,16 @@
+﻿using kOS.Suffixed;
+using UnityEngine;
+
+namespace kOS.Module
+{
+    [KSPAddon(KSPAddon.Startup.AllGameScenes, false)]
+    public class SceneChangeCleaner : MonoBehaviour
+    {
+        public void Awake()
+        {
+            VesselTarget.ClearInstanceCache();
+            BodyTarget.ClearInstanceCache();
+            Destroy(this);
+        }
+    }
+}